refactor(dashboard): split harness-health.ts god file into 3 modules
853 lines → 3 focused modules:
- harness-health-state.ts (269): types, AsyncResource, SSE handler,
  timer management, data loading
- harness-health-sections.ts (369): helper functions, StatusPill,
  StatCard, GateChart, section components
- harness-health.ts (247): HarnessHealth main component only

Dependency: harness-health → state + sections, sections → state types.
